+++ yak-security/src/main/java/io/yak/framework/security/service/UserRoleService.java
@@ -4,17 +4,17 @@ import io.yak.framework.security.common.entity.UserRole;
 import java.util.List;
 
 public interface UserRoleService {
-  public List<Long> getUserIdListByRoleId(Integer var1);
+  public List<Long> getUserIdListByRoleId(Long var1);
 
-  public List<Long> getRoleIdListByUserId(Integer var1);
+  public List<Long> getRoleIdListByUserId(Long var1);
 
-  public void updateUserRoleByUserId(Integer var1, List<Long> var2);
+  public void updateUserRoleByUserId(Long var1, List<Long> var2);
 
-  public void updateUserRoleByRoleId(Integer var1, List<Long> var2);
+  public void updateUserRoleByRoleId(Long var1, List<Long> var2);
 
-  public int getUserRoleCountByRoleId(Integer var1);
+  public int getUserRoleCountByRoleId(Long var1);
 
-  public int deleteByUserIdOrRoleId(Integer var1, Integer var2);
+  public int deleteByUserIdOrRoleId(Long var1, Long var2);
 
   public List<UserRole> getByRoleIds(List<Long> var1);
